requirements.txt
asgiref==3.3.4
Django==3.0.5
djongo==1.3.4
Pillow==8.2.0
pymongo==3.11.3
python-slugify==4.0.1
pytz==2021.1
sqlparse==0.2.4
text-unidecode==1.3
